refactor. Main.js, event handler location change
- wallet 디렉토리 안에 있던 main을 js 전체 main으로 변경함.
- 지갑 클릭 이벤트 영역을 model 영역으로 이동함.
